+++ src-tauri/src/ai_service/llm/providers/mod.rs
@@ -1,7 +1,6 @@
-mod codex;
 mod genai_provider;
 mod kimi_code;
 
-pub use codex::CodexProvider;
+pub use crate::ai_service::llm::codex::provider::CodexProvider;
 pub use genai_provider::GenaiProvider;
 pub use kimi_code::KimiCodeProvider;
